Add deck state
