+++ Website/data/locales/ja.json
@@ -100,6 +100,7 @@
 		"amountTaken": "{#CURRENT}/{#TOTAL} Taken",
 		"finished": "Draft finished. You can start the game now.",
 		"startGame": "Start Game",
+		"waitingForOpponent": "Waiting for Opponent...",
 		"deckName": "ドラフトデッキ",
 		"yourDeck": "自分のデッキ",
 		"opponentDeck": "相手のデッキ"
