+++ Website/index.html
@@ -477,7 +477,6 @@
 		<script src="scripts/ui.js"></script>
 		<script src="scripts/websocket.js"></script>
 		<script src="scripts/hotkeys.js"></script>
-		<script src="scripts/gameloop.js"></script>
 		<script src="scripts/cardActions.js"></script>
 	</body>
 </html>
