+++ Website/modules/gameState.js
@@ -1,6 +1,6 @@
-// This module exports the Gamestate class which serves as the base class for any distinct state/mode the game can be in.
+// This module exports the GameState class which serves as the base class for any distinct state/mode the game can be in.
 
-export class Gamestate {
+export class GameState {
 	receiveMessage(command, message) {
 		return false;
 	}
