+++ Website/scripts/variables.js
@@ -36,12 +36,16 @@ let ctrlHeld = false;
 let altHeld = false;
 
 let officialDecks = [];
+fetch("data/deckList.json")
+.then(response => response.json())
+.then(decks => {
+	officialDecks = decks;
+});
+
 let currentDeckList = "default";
 
 // decks and partner choices for both players
-let loadedDeck = null;
 let loadedPartner = null;
-let opponentDeck = null;
 let opponentPartner = null;
 
 let youAre = null; // Whether this client is player 0 or player 1. (Mainly for draft games and partner selection, as far as the board is concerned, the local player is always player 1.)
@@ -50,11 +54,3 @@ let canGrab = true; //whether or not cards can be grabbed. (only used when dropp
 let heldCard = null; // what card is currently being dragged
 let opponentHeldCard = null; // what card is currently being dragged by the opponent
 let opponentName = null; // The opponent's display name
-
-// local player and opponent cursor positions
-let myCursorX = 0;
-let myCursorY = 0;
-let oppCursorX = 0;
-let oppCursorY = 0;
-let oppCursorTargetX = 0;
-let oppCursorTargetY = 0;
